lead-agent/requirements.txt
anthropic>=0.39.0
google-cloud-firestore>=2.19.0
apscheduler>=3.10.4
python-dotenv>=1.0.1
aiohttp>=3.10.0
